requirements.txt
Flask
tensorflow
Pillow
requests
streamlit
gunicorn
